Make libconfig requirement optional
libconfig is a required to build libvirt-cim since r1104, but the
package is not included in RHEL 5, which we still support. This patch
enables libvirt-cim to be built even if libconfig is not found.

Note that it will not be possible for users of that distro to configure
libvirt-cim to operate in read-only mode.

Signed-off-by: Eduardo Lima (Etrunko) <eblima@br.ibm.com>