libxkutil: Handle vnc password when retrieving domain XML
Whenever a call for ModifyResourceSettings was issued, the XML for the
domain was requested, parsed, the property had its value changed and then
XML was updated with the new value.

It occurs that we need to specify the VIR_DOMAIN_XML_SECURE flag to retrieve
full domain info, including sensitive fields, such as passwords. This patch
fixes the problem for vnc password, which was not handled in the XML parsing
code.

Signed-off-by: Eduardo Lima (Etrunko) <eblima@br.ibm.com>