Redirect stdout & stderr when run as out-of-process provider
When a provider is loaded in-process by Pegasus, messages sent to
stdout and stderr are suppressed. However, when loaded out-of-process,
those same messages appear on the console. This happens excessively when
libvirt-cim LXC_ and XEN_ classes are enumerated and libvirt sends
"libvir: error : no connection driver available for..." messages to stderr.
In development, this can sometimes prevent console usage. Until a fix is
made to Pegasus, we'll suppress console output during provider initialization
when run in out-of-process mode.

Signed-off-by: Chip Vincent <cvincent@us.ibm.com>