device_parsing: Use default values for vnc graphics device
This patch fixes the behavior where libvirt-cim loses the graphics device
description after a call to ModifyResourceSettings method. Actually it has
nothing to do with the fact that the domain is running or not. What happens is
that if somehow we can't read either 'listen' or 'port' attributes, the
function will fail and return immediately, skipping the device inclusion.

The default values are based on the ones found in the default_graphics_device
function in src/Virt_VirtualSystemManagementService.c.

Signed-off-by: Eduardo Lima (Etrunko) <eblima@br.ibm.com>