Avoid connection to libvirt if previous attempt fails
This is a workaround to avoid libvirt flooding error messages in syslog. This
happens often if a client submits queries for CIM_ superclasses, which then will
translate to a query for each registered class. In our case KVM_, LXC_ and XEN_.

Ideally, there should be a way to ask libvirt if a given URI or hypervisor is
enabled/supported. A patch for that feature is on the works, and as soon as it
is integrated to libvirt tree this feature will be updated.

Signed-off-by: Eduardo Lima (Etrunko) <eblima@br.ibm.com>