Fix qos bug in primordial Network/0 pool
For some reason, when retreiving at SettingsDefinesCapabilites for only the
primordial network pool 'Network/0' the qos code is called for the
'Minimum' RASDs (?!), which because the associated ID is a constant
causes the string substitution code for renaming the qos 'Point/...' RASDs to
crash. This patch fixes this qos code so that the InstanceID string is
only manipulated for Point RASDs.

Signed-off-by: Gareth S. Bestor <bestor@us.ibm.com>