Pool Refresh Patch2
This patches earlier patch (from Sharad) that added pool refresh function to
libvirt-cim. Previously, a pool refresh would be initiated whenever the pool's
CIM instance was retreived; this turns out to be too expensive for some
consumers. This patch removes that, and instead makes pool refresh an
explicit CIM operation that a CIM client can request on the
KVM_ResourcePoolConfigurationService (eg when pool resources are added or
removed or changes outside of CIM). The previous patch also initiated a
refresh whenever libvirt-cim was used to add or remove a pool resource; this
fuction has been kept, because it makes sense for the CIM pool to immediately
reflect any changes when resources are added/removed via CIM.
The new extrinsic method is called RefreshResourcesInPool[<pool>]

Signed-off-by: Gareth S. Bestor <bestor@us.ibm.com>