Makefile.am: Remove the $(top_srcdir) from subst command
During the postinstall and preuninstall phases various variables are
modified to build up the list of mofs to be installed.  The generated
output had ".//usr/local/share/libvirt-cim/*" which caused issues finding
files. This is a followup to commit '22022870' which changed the paths using
to schema for each of the variables.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Reviewed-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Signed-off-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>