register: Adjust the chatter output
Caused error during make postinstall. Rather than print the $PROVIDERMODULES
it attempted to execute the second one in the list.

Added extra output just prior to CIMMOF commands to show which namespace
is being modified from which directory. There appears to be some sort of
issue registering some mofs as the following message is generated numerous
times:

Warning: the instance already exists.
In this implementation, that means it cannot be changed.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Reviewed-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>
Signed-off-by: Wenchao Xia <xiawenc@linux.vnet.ibm.com>