Coverity: Resolve NO_EFFECT - set_proc_rasd_params()
143             if (domain_online(dom))
144                     count = domain_vcpu_count(dom);
145             else
146                     count = dev->dev.vcpu.quantity;
147

(1) Event unsigned_compare:
    This greater-than-or-equal-to-zero comparison of an unsigned value
    is always true. "count >= 0UL".

148             if (count >= 0)

Resolve by adjusting logic.  Problem was that the active count is returned
as an int with an error value of -1, while the quantity value is guaranteed
to be 1 or more (see parse_vcpu_device() processing).  So initialize count to
zero, then only set the property if count > 0. Setting count of the active
condition requires a local "active_count" and checking that to be > 0 before
blindly setting it to count.  Imagine 0xfffffffffffffff vcpu's!