libxkutil: Provide easy access to the libvirt capabilities
Introspecting the libvirt capabilities and creating an internal capabilities
data structure. Methods are provided for retrieving default values regarding
architecture, machine and emulator for easy of use in the provider code.

Changed the KVM detection to use the capabilities instead of unique
XML parsing.

Further, xml_parse_test was extendend to display hypervisor capabilities
and defaults.

Signed-off-by: Boris Fiuczynski <fiuczy@linux.vnet.ibm.com>
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>