build: Don't use /bin/sh unconditionally
provider-register.sh implicitly assumes that the shell has bash-like
capabilities. On systems like Ubuntu this is not the case leading
to a make postinstall failure when used with Pegasus.
Changing the script shebang to /bin/bash to make this explicit.
Further replace occurrences of sh to $(SHELL) in the Makefile
invocations of provider-register.sh.

Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>