RASD: Schema and Provider Support for Console RASDs
Representing console devices was very limited with the previous
approach using a special graphical display. For instance it
was not possible to define different target types (serial, virtio,
sclp) nor to exploit all source types (pty, file, tcp, ...)
available in libvirt.

With the following new RASD classes it is possible to define
real console resources:
    KVM_ConsoleResouceAllocationSettingData
    Xen_ConsoleResouceAllocationSettingData
    LXC_ConsoleResouceAllocationSettingData

The ResourceType of the returned instances is is '1' (Other) and the
OtherResourceType is 'console'.

Implemented CIM operations:
   enumerate instances
   enumerate instance names
   get instance

Signed-off-by: Thilo Boehm <tboehm@linux.vnet.ibm.com>
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>
Reviewed-by: Boris Fiuczynski <fiuczy@linux.vnet.ibm.com>