libxkutil: Simplify XML handling of consoles
The attempt to avoid duplication of console definitions was a bit
too complicated and error prone. We move the generation of console
XML entirely to the new console code.
Further, it's incorrect to restrict PTY consoles being represented
as GraphicsRASDs to virtio only. At least serial must be allowed,
but it doesn't really hurt to map all PTYs to GraphicsRASD consoles.

Further, the parsing code returned two instances for serial consoles,
one from the /domain/devices/console and /domain/devices/serial
both representing the same device. This confuses cimtest and,
more problematic, will prevent DefineSystem to work, if a reference
VSSD with a serial console is passed in.

Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>
Reviewed-by: Daniel Hansel <daniel.hansel@linux.vnet.ibm.com>