build: Fix provider registration issues
This commit addresses an issue with provider registration during
RPM install or update.
The schema registration by wildcard doesn't take into consideration
that there are dependencies between the MOF files leading to
a partially populated repository and a not working libvirt-cim
provider.
Fixed by explicitly stating the mof and registration files in the
necessary order.

Further a minor false error message coming from the systemd service
detection was bound to cause irritation. This is suppressed now.

Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>