Use of root/interop instead of root/PG_InterOp
As of tog-pegasus 2.12.1-5 we can no longer use "root/PG_InterOp" namespace
for the CIM provider instead the use of "root/interop" is the preferred
mechanism.

This patch will adjust where libvirt-cim installs its classes to use the
"root/interop" namespace.

For more context, see:

    http://www.redhat.com/archives/libvirt-cim/2013-November/msg00083.html

and

    http://www.redhat.com/archives/libvirt-cim/2013-November/msg00008.html

This patch does not include the schema changes to FilterList although I
have a suspicion that a bug fix between 2.12.1-8 and 2.12.1-11 has resolved
the issue seen.  Cannot find a reference though.