Parse/Store controller XML tags
Parse and store the <controller> XML tags.  The generated DeviceID will
be a combination of "controller" + XML device "name" + XML Index "value".
This should be unique enough now and generates "controller:pci:0" or
"controller:usb:0" DeviceID's.

In the future, if support is added for the "<master>" XML child a new
mechanism will need to be put in place since one of those controllers is
designed to have the same name and index - what changes is the function
for the device.

Signed-off-by: John Ferlan <jferlan@redhat.com>